Middleware auth apply

+++ service/index.js
@@ -9,20 +9,24 @@ const getContactById = async (userId, contactId) => {
   return Contact.findOne({ owner: userId, _id: contactId });
 };
 
-const addContact = async (contact) => {
-  return Contact.create(contact);
+const addContact = async (userId, contact) => {
+  return Contact.create({ owner: userId }, contact);
 };
 
-const removeContact = async (contactId) => {
-  return Contact.findOneAndDelete({ _id: contactId });
+const removeContact = async (userId, contactId) => {
+  return Contact.findOneAndDelete({ owner: userId, _id: contactId });
 };
 
-const updateContact = async (contactId, body) => {
-  return Contact.findByIdAndUpdate({ _id: contactId }, body, { new: true });
+const updateContact = async (userId, contactId, body) => {
+  return Contact.findByIdAndUpdate({ owner: userId, _id: contactId }, body, {
+    new: true,
+  });
 };
 
-const updateStatusContact = async (contactId, body) => {
-  return Contact.findByIdAndUpdate({ _id: contactId }, body, { new: true });
+const updateStatusContact = async (userId, contactId, body) => {
+  return Contact.findByIdAndUpdate({ owner: userId, _id: contactId }, body, {
+    new: true,
+  });
 };
 
 const addUser = async (user) => {
